dom/base/test/test_bug590812.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for XML pretty printing, bug 590812</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/WindowSnapshot.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body onload="runTest()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=590812">Mozilla Bug 590812</a>
<p id="display"></p>
<iframe src="http://noxul.example.com/tests/dom/base/test/file_bug590812.xml"></iframe>
<iframe src="file_bug590812.xml"></iframe>
<iframe src="file_bug590812-ref.xhtml"></iframe>
<pre id="test">
<script class="testbody" type="application/javascript">

SimpleTest.waitForExplicitFinish();

async function runTest() {
  // FIXME(Fission): The load event fires before cross-origin iframes have
  // loaded (bug 1559841).
  if (SpecialPowers.useRemoteSubframes) {
    for (let i = 0; i < 100; i++) {
      await new Promise(resolve => setTimeout(resolve, 0));
    }
  }

  let sNoXUL = await snapshotWindow(window.frames[0], false);
  let sWithXUL = await snapshotWindow(window.frames[1], false);
  let sRef = await snapshotWindow(window.frames[2], false);
  let res;
  ok(compareSnapshots(sNoXUL, sRef, true)[0],
     "noxul domain same as ref");
  ok(compareSnapshots(sWithXUL, sRef, true)[0],
     "xul supporting domain same as ref");

  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
